math test

+++ util/arbmath/math.go
@@ -320,7 +320,7 @@ func SaturatingMul[T Signed](a, b T) T {
 // SaturatingCast cast an unsigned integer to a signed one, clipping to [0, S::MAX]
 func SaturatingCast[S Signed, T Unsigned](value T) S {
 	tBig := unsafe.Sizeof(T(0)) >= unsafe.Sizeof(S(0))
-	bits := 8 * unsafe.Sizeof(S(0))
+	bits := uint64(8 * unsafe.Sizeof(S(0)))
 	sMax := T(1<<bits-1) >> 1
 	if tBig && value > sMax {
 		return S(sMax)
